Finalize Game Logic
Main function playGame() starts at page load. Inside the function define are playRound, getHumanChoice, getComputerChoice which are at the same levels.

The main loop starts using while  until  any of the score is equal to 5.  Return statements are used if at least one of players has reached 5 to break the loop and with the announcement of the Champion.

Modify HTML to put a title in the Body with rules as  heading followed by the list.